+++ app/Views/admin/attendance/pending-absences.php
@@ -14,7 +14,7 @@
             </div>
             <div class="col-sm-6">
                 <ol class="breadcrumb float-sm-right">
-                    <li class="breadcrumb-item"><a href="/panel/dashboard">Dashboard</a></li>
+                    <li class="breadcrumb-item"><a href="/panel/dashboard">Inicio</a></li>
                     <li class="breadcrumb-item"><a href="/panel/attendance">Marcaciones</a></li>
                     <li class="breadcrumb-item active">Ausencias Pendientes</li>
                 </ol>